refactor: test case separate
